:twisted_rightwards_arrows: ADD. feature/#5-chatbot branch merge
