+++ src/main/java/com/tantanmen/carbofootprint/domain/member/enums/MemberRoles.java
@@ -1,0 +1,5 @@
+package com.tantanmen.carbofootprint.domain.member.enums;
+
+public enum MemberRoles {
+    ADMIN, CLIENT
+}
