+++ src/main/java/com/tantanmen/carbofootprint/domain/member/repository/AuthorityRepository.java
@@ -1,6 +1,6 @@
-package com.tantanmen.carbofootprint.domain.user.repository;
+package com.tantanmen.carbofootprint.domain.member.repository;
 
-import com.tantanmen.carbofootprint.domain.user.entity.Authority;
+import com.tantanmen.carbofootprint.domain.member.entity.Authority;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AuthorityRepository extends JpaRepository<Authority, Long> {
