+++ src/main/java/com/tantanmen/carbofootprint/domain/member/web/controller/MemberController.java
@@ -1,11 +1,11 @@
-package com.tantanmen.carbofootprint.domain.user.web.controller;
+package com.tantanmen.carbofootprint.domain.member.web.controller;
 
-import com.tantanmen.carbofootprint.domain.user.repository.MemberRepository;
-import com.tantanmen.carbofootprint.domain.user.service.MemberService;
-import com.tantanmen.carbofootprint.domain.user.web.dto.LoginRequestDto;
-import com.tantanmen.carbofootprint.domain.user.web.dto.LoginResponseDto;
-import com.tantanmen.carbofootprint.domain.user.web.dto.SignUpResponseDto;
-import com.tantanmen.carbofootprint.domain.user.web.dto.SignUpRequestDto;
+import com.tantanmen.carbofootprint.domain.member.repository.MemberRepository;
+import com.tantanmen.carbofootprint.domain.member.service.MemberService;
+import com.tantanmen.carbofootprint.domain.member.web.dto.LoginRequestDto;
+import com.tantanmen.carbofootprint.domain.member.web.dto.LoginResponseDto;
+import com.tantanmen.carbofootprint.domain.member.web.dto.SignUpResponseDto;
+import com.tantanmen.carbofootprint.domain.member.web.dto.SignUpRequestDto;
 import com.tantanmen.carbofootprint.global.entity.response.CustomApiResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
