+++ src/main/java/com/tantanmen/carbofootprint/domain/member/web/dto/SignUpResponseDto.java
@@ -1,4 +1,4 @@
-package com.tantanmen.carbofootprint.domain.user.web.dto;
+package com.tantanmen.carbofootprint.domain.member.web.dto;
 
 import lombok.Builder;
 import lombok.Getter;
