--- src/main/java/com/tantanmen/carbofootprint/domain/user/entity/Authority.java
@@ -1,31 +1,0 @@
-package com.tantanmen.carbofootprint.domain.user.entity;
-
-import com.tantanmen.carbofootprint.domain.user.enums.MemberRoles;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Authority {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long authority_id;
-
-    @Enumerated(EnumType.STRING)
-    private MemberRoles authorityType;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    public void setMember(Member member) {
-        this.member = member;
-    }
-}
