--- src/main/java/com/tantanmen/carbofootprint/domain/user/entity/Member.java
@@ -1,36 +1,0 @@
-package com.tantanmen.carbofootprint.domain.user.entity;
-
-import com.tantanmen.carbofootprint.global.entity.BaseEntity;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Builder
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class Member extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long memberId;
-
-    private String loginId;
-
-    private String password;
-
-    @OneToMany(mappedBy = "member", fetch = FetchType.EAGER, cascade = CascadeType.ALL)
-    private List<Authority> authorities = new ArrayList<>();
-
-    public void addRole(Authority authority) {
-        authorities.add(authority);
-        authority.setMember(this);
-    }
-
-}
