--- src/main/java/com/tantanmen/carbofootprint/domain/user/enums/MemberRoles.java
@@ -1,5 +1,0 @@
-package com.tantanmen.carbofootprint.domain.user.enums;
-
-public enum MemberRoles {
-    ADMIN, CLIENT
-}
