+++ src/main/java/com/tantanmen/carbofootprint/global/security/service/CustomUserDetails.java
@@ -1,6 +1,6 @@
 package com.tantanmen.carbofootprint.global.security.service;
 
-import com.tantanmen.carbofootprint.domain.user.entity.Member;
+import com.tantanmen.carbofootprint.domain.member.entity.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
